service/accent-service/pyproject.toml
[project]
name = "accent-service"
version = "0.1.0"
description = ""
authors = [{ name = "Ryan Clark", email = "ryanc@accentservices.com" }]
requires-python = ">=3.9"
dependencies = ["dbus-python>=1.2.16", "accent-manage-db"]

[project.scripts]
accent-service-python = "accent_service.main:main"

[project.entry-points]

[dependency-groups]
dev = ["pytest>=8.3.5"]

[tool.uv.sources]
accent-manage-db = { workspace = true }

[tool.hatch.build.targets.sdist]
include = ["accent_service", "bin/*"]

[tool.hatch.build.targets.wheel]
include = ["accent_service", "bin/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
